+++ src/hypothesis_openapi/_serialize.py
@@ -1,4 +1,4 @@
-from dataclasses import is_dataclass, fields
+from dataclasses import fields, is_dataclass
 
 from ._types import Missing